% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{predict_countries}
\alias{predict_countries}
\alias{eha_countries}
\alias{ehp_countries}
\title{This function returns the names of PREDICT-2 countries in the EIDITH database.}
\usage{
predict_countries()

eha_countries()

ehp_countries()
}
\description{
This function returns the names of PREDICT-2 countries in the EIDITH database.

This function returns the names of EHA PREDICT-2 countries in the EIDITH database.

This function returns the names of Ebola Host Project PREDICT-2 countries in the EIDITH database.
}
