% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ed2_expand_long}
\alias{ed2_expand_long}
\title{Expand multiple-response fields into long data form}
\usage{
ed2_expand_long(df, col, other_details = TRUE)
}
\arguments{
\item{df}{The data frame containing the field to expand}

\item{col}{The column name to split into long form}

\item{other_details}{Whether other category should be consolidated and a new column with other details added. Defaults to TRUE.}
}
\value{
A long form dataframe with \code{col} field separated into repeated entries.
}
\description{
Expand multiple-response fields into long data form
}
\examples{
\dontrun{
humans <- ed2_human()
humans_reduced <- dplyr::select(humans, event_name, participant_id, travel_reason)
humans_reduced_long <- ed2_expand_long(humans_reduced, travel_reason)
}
}
