% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ed2_expand_wide}
\alias{ed2_expand_wide}
\title{Expand multiple-response fields into wide data form}
\usage{
ed2_expand_wide(df, col, clean_names = TRUE)
}
\arguments{
\item{df}{The data frame containing the field to expand}

\item{col}{The column name to split into wide form}

\item{clean_names}{Whether new wide field names should be cleaned. Defaults to TRUE.}
}
\value{
A wide form dataframe with \code{col} field separated into different columns and filled with TRUE or FALSE values.
}
\description{
Expand multiple-response fields into wide data form
}
\examples{
\dontrun{
humans <- ed2_human()
humans_reduced <- dplyr::select(humans, event_name, participant_id, travel_reason)
humans_reduced_wide <- ed2_expand_wide(humans_reduced, travel_reason)
}
}
