% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{ed2_metadata}
\alias{ed2_metadata}
\title{EIDITH PREDICT-2 table metadata}
\usage{
ed2_metadata()
}
\description{
Descriptions of fields in the EIDITH database.  Running \code{ed2_metadata()} will
return a data frame with this information.
}
\details{
More information can be found at \href{https://eidith.org/Resources.aspx}{the EIDITH resources page}.

\if{html}{
  \Sexpr[echo=FALSE, results=rd, stage=build]{
  in_pkgdown <- any(grepl("as_html.tag_Sexpr", sapply(sys.calls(), function(a) paste(deparse(a), collapse = "\n"))))
    if(in_pkgdown) {
      mytext <- c('In RStudio, this help file includes a searchable table of values.')
    } else {
      mytext <- eidith::rd_datatable(eidith::ed2_metadata())
    }
    mytext
  }
}

\if{text,latex}{The HTML version of this help file includes a searchable table of values.}
}
