% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{ed2_table}
\alias{ed2_table}
\alias{ed2_events}
\alias{ed2_animals}
\alias{ed2_specimens}
\alias{ed2_animal_production}
\alias{ed2_crop_production}
\alias{ed2_zoo_sanctuary}
\alias{ed2_wildlife_restaurant}
\alias{ed2_natural_areas}
\alias{ed2_market_value_chain}
\alias{ed2_human}
\alias{ed2_tests}
\alias{ed2_test_serology}
\alias{ed2_test_interpreted}
\alias{ed2_human_animal_production}
\alias{ed2_extractive_industry}
\alias{ed2_dwellings}
\alias{ed2_human_crop_production}
\alias{ed2_human_extractive_industry}
\alias{ed2_human_hospital_worker}
\alias{ed2_human_hunter}
\alias{ed2_human_market}
\alias{ed2_human_restaurant}
\alias{ed2_human_sick_person}
\alias{ed2_human_temporary_settlements}
\alias{ed2_human_zoo}
\title{Load PREDICT-2 EIDITH data from the local database}
\usage{
ed2_events(...)

ed2_animals(...)

ed2_specimens(...)

ed2_animal_production(...)

ed2_crop_production(...)

ed2_zoo_sanctuary(...)

ed2_wildlife_restaurant(...)

ed2_natural_areas(...)

ed2_market_value_chain(...)

ed2_human(...)

ed2_tests(...)

ed2_test_serology(...)

ed2_test_interpreted(...)

ed2_human_animal_production(...)

ed2_extractive_industry(...)

ed2_dwellings(...)

ed2_human_crop_production(...)

ed2_human_extractive_industry(...)

ed2_human_hospital_worker(...)

ed2_human_hunter(...)

ed2_human_market(...)

ed2_human_restaurant(...)

ed2_human_sick_person(...)

ed2_human_temporary_settlements(...)

ed2_human_zoo(...)
}
\arguments{
\item{...}{arguments passed to \code{\link[dplyr:filter]{dplyr::filter()}} to subset data}
}
\value{
a \link[tibble:tibble]{tibble}-style data frame.
}
\description{
These functions load data from the locally-stored SQLite database of downloaded
and cleaned EIDITH data. \code{ed_table} is a general function, and \code{ed_table_}
it's \href{https://cran.r-project.org/web/packages/dplyr/vignettes/nse.html}{standard evaluation}
equivalent, useful for programming.  The other functions are convenience aliases
for the individual tables. Functions beginning with the \code{ed} prefix refer to
PREDICT-1 data tables, while \code{ed2} functions refer to PREDICT-2 data tables.
}
\details{
These functions take \code{\link[dplyr:filter]{dplyr::filter()}} arguments to sub-set the data.  Using
these, the data is subsetted via SQL \emph{before} it is loaded into memory.
For large tables, such as the \emph{tests} table, this is useful for reducing the memory footprint of your R session.

Note that subsetting in SQL is more limited:
\itemize{
\item Use \code{0} or \code{1} instead of \code{TRUE} or \code{FALSE}
\item Dates are stored as character strings, but as they are in YYYY-MM-DD
format, filtering such as \code{event_date > "2014-01-01"} still works.
}
}
