% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{ed_auth}
\alias{ed_auth}
\title{EIDITH logon credentials}
\description{
Any function that downloads data from EIDITH, such as \code{\link[=ed_db_download]{ed_db_download()}} or
the \link[=ed_get]{raw download functions} requires login credentials.  These
functions will ask for a username and password.  To get an EIDITH account
with data access permissions, contact technology@eidith.org.
}
\details{
For frequent or programmatic use, we recommend caching your logon credentials
as environment variables.  \strong{eidith} functions will automatically search
for your credentials in the \code{EIDITH_USERNAME} and \code{EIDITH_PASSWORD} environment
variables.

To cache your credentials across multiple R sessions, save them in your
personal \code{.Renviron} file. You can also use the \code{usethis} package to easily access
and edit your \code{.Renviron} file:\preformatted{install.packages("usethis")
usethis::edit_r_environ()
}

This should open up your \code{.Renviron} file in a new tab, and you add the lines:\preformatted{EIDITH_USERNAME=your_username
EIDITH_PASSWORD=your_password
}

Save this file, restart R, and you will be automatically logged in to download data from EIDITH
when working on this computer.
}
