% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact.R
\name{ed_contact}
\alias{ed_contact}
\alias{contact}
\alias{support}
\alias{ed_support}
\title{Contacts for support}
\description{
The \strong{eidith} R package is maintained by the PREDICT Modeling & Analytics
team at EcoHealth Alliance.  The EIDITH database is maintained by the PREDICT
Information Mananagement team at metabiota.

For bug reports and issues with the \strong{eidith} R package, please file an
issue at with EcoHealth Alliance M&A at \url{https://github.com/ecohealthalliance/eidith/issues} and we will
repond promptly.  You may also contact Noam Ross (\href{mailto:ross@ecohealthalliance.org}{ross@ecoehalthallinace.org})
and Cale Basaraba (\href{mailto:basaraba@ecohealthalliance.org}{basaraba@ecoehalthalliance.org}) directly.

For issues related to your EIDITH login credentials or incorrect data,
contact \href{mailto:technology@eidith.org}{technology@eidith.org}.

Download failures may either be due to issues at the database or package level.
If you are unsure which team is the right one to address your issue, we suggest
you contact EcoHealth Alliance M&A first.
}
