% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_checks.R
\name{ed_db_check_status}
\alias{ed_db_check_status}
\title{Get the status of the locally stored EIDITH database}
\usage{
ed_db_check_status(path = NULL, inter = T)
}
\arguments{
\item{path}{if provided, the filename of the sqlite database to check. By default,
the function checks the status of the internal database or that with the global option \code{"ed_sql_path"}.}

\item{inter}{whether function is run interactively}
}
\value{
A message regarding the status of current EIDITH database, or a prompt allowing the user
to download missing / corrupt tables
}
\description{
This function provides:
}
\seealso{
\code{\link[=ed_db_download]{ed_db_download()}}, \code{\link[=ed_db_updates]{ed_db_updates()}}, \code{\link[=ed_db_export]{ed_db_export()}}
}
