% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_checks.R
\name{ed_db_detailed_status}
\alias{ed_db_detailed_status}
\title{Obtain and print status of all local EIDITH tables}
\usage{
ed_db_detailed_status()
}
\description{
This function evaluates and prints a detailed banner of local EIDITH PREDICT-1
and PREDICT-2 table statuses. This includes whether they are currently present,
their last download date, and the PREDICT-1 and PREDICT-2 countries that
are available in the local database.
}
