% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{ed_db_download}
\alias{ed_db_download}
\title{Download EIDITH data to local storage}
\usage{
ed_db_download(p1_tables = p1_api_endpoints(),
  p2_tables = p2_api_endpoints(), country = NULL, p1_data = FALSE,
  verbose = interactive())
}
\arguments{
\item{p1_tables}{Which PREDICT-1 tables to download}

\item{p2_tables}{which PREDICT-2 tables to download}

\item{verbose}{Show messages while in progress?}
}
\description{
This function fetches all available tables from the EIDITH database and
caches the data locally for further use with the \link[=ed_table]{table functions}.
The database persists between R sessions, so you only need to run this when
you want to update data.  You should also run it when updating to new
versions of the \strong{eidith} package.
}
\details{
This function does a few things:
\itemize{
\item Fetches all tables from both EIDITH PREDICT-1 and PREDICT-2 databases.
\item Cleans and pre-processes the data with \code{\link[=ed_process]{ed_process()}}.
\item Stores the data in an SQLite database in your \link[rappdirs:user_data_dir]{user data directory}.
}

Each time this function is run it checks for updated metadata about database tables,
then downloads each table into a temporary database and makes sure that the
structure of these tables is as expected. Only well-formed tables will be copied
into the user's permanent database. If a downloaded table is missing fields, has
extra fields, or is otherwise corrupted, the user will receive an error message.
In the future, we will allow for updating only with new and changed records.

The function will prompt for username and password unless you have \link[=ed_auth]{cached your credentials}.
}
\seealso{
\code{\link[=ed_db_status]{ed_db_status()}}, \code{\link[=ed_db_updates]{ed_db_updates()}}, \code{\link[=ed_db_export]{ed_db_export()}}
}
