% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{ed_db_export}
\alias{ed_db_export}
\title{Export the local EIDITH database to a file}
\usage{
ed_db_export(filename, ...)
}
\arguments{
\item{filename}{The filename to export to. We suggest something ending in \code{.sqlite}.}

\item{...}{Other options passed to \code{\link[=file.copy]{file.copy()}}}
}
\description{
This function allows you to export the local EIDITH database to a file that
can then be used by others.  The database is in \href{https://sqlite.org/}{SQLite} format.
}
\examples{
\dontrun{
  #Here's an example of how to export and then use the exported database

  ed_db_export("mydb.sqlite")
  options(ed_sql_path = "mydb.sqlite") # This switches to working with the exported database
  ed_db_status()  #get status of the current (exported) database
}
}
\seealso{
\code{\link[=ed_db_status]{ed_db_status()}}, \code{\link[=ed_db_updates]{ed_db_updates()}}, \code{\link[=ed_db_export]{ed_db_export()}}
}
