% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{ed_db_status}
\alias{ed_db_status}
\title{Get the status of the locally stored EIDITH database}
\usage{
ed_db_status(path = NULL)
}
\arguments{
\item{path}{if provided, the filename of the sqlite database to check. By default,
the function checks the status of the internal database or that with the global option \code{"ed_sql_path"}.}
}
\value{
A list of database status information, pretty-printed.
}
\description{
This function provides:
\itemize{
\item Countries in the database
\item Size of each table in the database
\item The most recent download date
\item The last-updated records in local database
}
}
\seealso{
\code{\link[=ed_db_download]{ed_db_download()}}, \code{\link[=ed_db_updates]{ed_db_updates()}}, \code{\link[=ed_db_export]{ed_db_export()}}
}
