% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{ed_db_updates}
\alias{ed_db_updates}
\title{Check the online EIDITH databse for updates since your last download.}
\usage{
ed_db_updates(path = NULL)
}
\arguments{
\item{path}{if provided, the filename of the sqlite database to check. By default,
the function checks the status of the internal database or that with the global option \code{"ed_sql_path"}.}
}
\value{
A named vector, showing TRUE for tables that have been updated.  A message is also printed.
}
\description{
This function checks to see if the EIDITH online database has been updated
since the last time \code{\link[=ed_db_download]{ed_db_download()}} was run. If it has, you may run
\code{\link[=ed_db_download]{ed_db_download()}} to get the latest data.
}
\seealso{
\code{\link[=ed_db_status]{ed_db_status()}}, \code{\link[=ed_db_download]{ed_db_download()}},  \code{\link[=ed_db_export]{ed_db_export()}}
}
