% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta.R
\name{ed_fasta}
\alias{ed_fasta}
\alias{fasta}
\alias{ed_fasta_}
\alias{ed_tests_report}
\title{Export FASTA files}
\usage{
ed_fasta(.data, file = "", sequence = "sequence", ...)

ed_fasta_(.data, file = "", sequence = "sequence", ..., .dots)

ed_tests_report()
}
\arguments{
\item{.data}{The data frame}

\item{file}{name of file to write to. If "" (default), the output is written to console.}

\item{sequence}{the column name in the data frame that contains the sequence.  Default is "sequence".}

\item{...}{Additional columns to include as metadata, using \code{\link[dplyr:select]{dplyr::select()}} syntax.  If nothing is included, all columns are used.}

\item{.dots}{Columns specified using standard evaluation.}
}
\value{
The FASTA file is written to a file or the console.  Also, a character vector of each FASTA record is returned invisibly.  \code{ed_tests_report()} returns a data frame.
}
\description{
\code{ed_fasta()} and \code{ed_fasta_()} convert a data frame to a FASTA file.  \code{ed_tests_report()} is a standard report returning a data frame of all uninterpreted test sequences, along with a set of metadata, which is commonly printed to FASTA
}
\examples{
\dontrun{
 ed_fasta(ed_tests_report())
 }
}
