% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{ed_metadata}
\alias{ed_metadata}
\alias{ed_taxagroups}
\title{EIDITH PREDICT-1 table metadata}
\usage{
ed_metadata()

ed_taxagroups()
}
\description{
Descriptions of fields in the EIDITH database.  Running \code{ed_metadata()} will
return a data frame with this information. \code{ed_taxagroups} returns a table
of taxagroup-to-order mapping.
}
\details{
More information can be found at \href{https://eidith.org/Resources.aspx}{the EIDITH resources page}.

\if{html}{
  \Sexpr[echo=FALSE, results=rd, stage=build]{
  in_pkgdown <- any(grepl("as_html.tag_Sexpr", sapply(sys.calls(), function(a) paste(deparse(a), collapse = "\n"))))
    if(in_pkgdown) {
      mytext <- c('In RStudio, this help file includes a searchable table of values.')
    } else {
      mytext <- eidith::rd_datatable(eidith::ed_metadata())
    }
    mytext
  }
}

\if{text,latex}{The HTML version of this help file includes a searchable table of values.}
}
