% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_data_calls.R
\name{ed_table_mock_}
\alias{ed_table_mock_}
\alias{ed_mock}
\alias{ed_table_mock}
\alias{ed_mock}
\alias{ed_events_mock}
\alias{ed_mock}
\alias{ed_animals_mock}
\alias{ed_mock}
\alias{ed_specimens_mock}
\alias{ed_mock}
\alias{ed_tests_mock}
\alias{ed_mock}
\alias{ed_viruses_mock}
\alias{ed_mock}
\alias{ed_testspecimen_mock}
\alias{ed_mock}
\title{Load mock EIDITH data for tutorials}
\usage{
ed_table_mock_(mock_table, ..., .dots, raw = FALSE)

ed_table_mock(mock_table, ...)

ed_events_mock(...)

ed_animals_mock(...)

ed_specimens_mock(...)

ed_tests_mock(...)

ed_viruses_mock(...)

ed_testspecimen_mock(...)
}
\arguments{
\item{mock_table}{one of the EIDITH database tables. One of "events", "animals",
"specimens", "tests", "viruses", or "test_specimen_ids".}

\item{...}{arguments passed to \code{\link[dplyr:filter]{dplyr::filter()}} to subset data}

\item{.dots}{standard-evaluation versions of subsetting arguments}

\item{raw}{Whether to return the unprocessed form of the mock data.  This
is equivalent to calling the \code{\link[=ed_get]{ed_get()}} functions with \code{postprocess=FALSE}}
}
\value{
a \link[tibble:tibble]{tibble}-style data frame.
}
\description{
These functions load mock versions of the EIDITH data tables to be used
for demonstrations, tutorials, and practice.  They are drop-in replacements
for the \code{\link[=ed_table]{ed_table()}} functions.
}
