% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{ed_postprocess}
\alias{ed_postprocess}
\title{Standard post-processing of EIDITH exports}
\usage{
ed_postprocess(data, endpoint)
}
\arguments{
\item{data}{The data as exported from EIDITH and imported via}

\item{endpoint}{The name of the endpoint or data table: one of "Event",
"Animal",  "Specimen", "Test", or "Virus".}
}
\description{
This function takes raw EIDITH data tables and does minor common
post-processing tasks.
}

