% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{ed_process}
\alias{ed_process}
\title{Standard post-processing of EIDITH exports}
\usage{
ed_process(dat, endpoint)
}
\arguments{
\item{dat}{The dat as exported from EIDITH and imported via}

\item{endpoint}{The name of the endpoint or dat table: one of "Event",
"Animal",  "Specimen", "Test", or "Virus".}
}
\description{
This function takes raw EIDITH dat tables and does minor common
post-processing tasks.
}

