% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eidith.R
\docType{package}
\name{eidith}
\alias{eidith}
\alias{eidith-package}
\title{eidith}
\description{
The \strong{eidith} R package provides programmatic access and analytical tools
for data from the \href{http://www.vetmed.ucdavis.edu/ohi/predict/}{PREDICT program},
housed at the \href{https://www.eidith.org/}{Emerging Infectious Disease Information Technology Hub}.
}
\details{
The \strong{eidith} package contains no data. To access data, you must be
\href{https://www.eidith.org/register.aspx}{registered} EIDITH user with data access privileges.
}
\section{Topics}{

\itemize{
\item \link[=ed_auth]{Authentication with EIDITH credentials}
\item Downloading EIDITH data to a \link[=ed_db_download]{local store} and \link[=ed_get]{direct API access}
\item \link[=ed_process]{Data cleaning}
\item \link[=ed_metadata]{Metadata for tables and fields in the EIDITH database}
}
}

\section{Contact}{

\itemize{
\item Package source code is found at \url{https://github.com/ecohealthalliance/eidith}
\item File bug reports and request features at \url{https://github.com/ecohealthalliance/eidith/issues}
\item Online version of this documentation found at \href{https://github.com/ecohealthalliance/eidith/issues}{https://ecohealthalliance.github.io/eidith/}
\item Questions related to your EIDITH credentials or incorrect data should be directed to \href{mailto:technology@eidith.org}{technology@eidith.org}.
\item More detail at \link[=ed_contact]{?ed_contact}.
}

To learn more about eidith, start with the vignettes:
\code{browseVignettes(package = "eidith")}
}

