% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{p1_api_endpoints}
\alias{p1_api_endpoints}
\alias{p2_api_endpoints}
\alias{p2_api_endpoints}
\alias{p1_api_endpoints}
\title{The names of the tables in the EIDITH database}
\usage{
p1_api_endpoints()

p2_api_endpoints()
}
\description{
The names of the tables in the EIDITH database
}
