% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{get_eidith_data}
\alias{ed_animals}
\alias{ed_events}
\alias{ed_specimens}
\alias{ed_tests}
\alias{ed_viruses}
\alias{get_eidith_data}
\title{Functions to access main EIDITH tables}
\usage{
get_eidith_data(endpoint, username = eidith_user(), password = eidith_pwd(),
  progress = interactive())

ed_events(username = eidith_user(), password = eidith_pwd())

ed_animals(username = eidith_user(), password = eidith_pwd())

ed_specimens(username = eidith_user(), password = eidith_pwd())

ed_tests(username = eidith_user(), password = eidith_pwd())

ed_viruses(username = eidith_user(), password = eidith_pwd())
}
\arguments{
\item{endpoint}{Which API endpoint to access, one of "Event", "Animal",
"Specimen", "Test", or "Virus". Each endpoint delivers one of these tables.
Convenience functions are provided for each.}

\item{username}{Your EIDITH username. Store it in your \code{.Renviron} file
as \code{EIDITH_USERNAME=XXXXX} for automatic access.}

\item{progress}{Show a progress bar?}

\item{username}{Your EIDITH password. Store it in your \code{.Renviron} file
as \code{EIDITH_PASSWORD=XXXXX} for automatic access.}
}
\value{
a \link[tibble]{tibble}-style data frame
}
\description{
Functions to access main EIDITH tables
}

