% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta.R
\name{ed_fasta}
\alias{ed_fasta}
\alias{fasta}
\title{Export FASTA files for viral interpretation}
\usage{
ed_fasta(
  .data,
  filename = "",
  sequence = "sequence",
  include_names = TRUE,
  ...
)
}
\arguments{
\item{.data}{The data frame}

\item{filename}{name of file to write to. If "" (default), the output is written to console.}

\item{sequence}{the column name in the data frame that contains the sequence.  Default is "sequence".}

\item{include_names}{Whether to include column names of sequence discriptors.  Default is TRUE.}

\item{...}{Additional columns to include as metadata, using \code{\link[dplyr:select]{dplyr::select()}} syntax.  If nothing is included, all columns are used.}
}
\value{
For \code{ed_fasta()}, \code{ed_fasta_group()}, and \code{ed_report_excel()}, the file(s) will be written to a file or the console (default).  Also, a character vector of each FASTA record is returned invisibly
\code{ed_tests_report()} returns a data frame.
}
\description{
\code{ed_fasta()} and \code{ed_fasta_group()} convert a data frame to a FASTA file.  \code{ed_tests_report()} is a standard report returning a data frame of all uninterpreted test sequences, along with a set of metadata, which is commonly printed to FASTA.
\code{ed_report_excel()} is a convenience function for creating excel-ready CSVs for reporting.  See the \href{https://ecohealthalliance.github.io/eidith/articles/fasta_walkthrough.html}{FASTA vignette} for more guidance.
}
\examples{
\dontrun{
 ed_fasta(ed_tests_report())
 }
}
