% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{ed_get}
\alias{ed_get}
\alias{ed_get_events}
\alias{ed_get_animals}
\alias{ed_get_specimens}
\alias{ed_get_tests}
\alias{ed_get_viruses}
\alias{ed_get_testspecimen}
\title{Functions to download EIDITH tables via API}
\usage{
ed_get(endpoint, verbose = interactive(), postprocess = TRUE,
  header_only = FALSE, lmdate_from = "2000-01-01", lmdate_to = Sys.Date()
  + 1, auth = NULL, ...)

ed_get_events(verbose = interactive(), postprocess = TRUE,
  header_only = FALSE, lmdate_from = "2000-01-01", lmdate_to = Sys.Date()
  + 1, ...)

ed_get_animals(verbose = interactive(), postprocess = TRUE,
  header_only = FALSE, lmdate_from = "2000-01-01", lmdate_to = Sys.Date()
  + 1, ...)

ed_get_specimens(verbose = interactive(), postprocess = TRUE,
  header_only = FALSE, lmdate_from = "2000-01-01", lmdate_to = Sys.Date()
  + 1, ...)

ed_get_tests(verbose = interactive(), postprocess = TRUE,
  header_only = FALSE, lmdate_from = "2000-01-01", lmdate_to = Sys.Date()
  + 1, ...)

ed_get_viruses(verbose = interactive(), postprocess = TRUE,
  header_only = FALSE, lmdate_from = "2000-01-01", lmdate_to = Sys.Date()
  + 1, ...)

ed_get_testspecimen(verbose = interactive(), postprocess = TRUE,
  header_only = FALSE, lmdate_from = "2000-01-01", lmdate_to = Sys.Date()
  + 1, ...)
}
\arguments{
\item{verbose}{Show a progress bar and other messages?}

\item{postprocess}{Should data be cleaned via \code{\link[=ed_process]{ed_process()}} or returned raw?}

\item{header_only}{Return only the table header.  Useful for checking if
API access works}

\item{lmdate_from}{filter records by earliest date, in YYYY-MM-DD format}

\item{lmdate_to}{filter records by latest date, in YYYY-MM-DD format (see details)}

\item{...}{additional arguments passed to \code{\link[httr:GET]{httr::GET()}}}
}
\value{
a \link[tibble:tibble]{tibble}-style data frame
}
\description{
These functions download data directly from the EIDITH API.  They require
\link[=ed_auth]{authorization}.  They can be useful for comparing local data
against data updated in the database, or processed vs. unprocessed data.
The \link[=ed_table]{table functions} load data from the local database instead,
and are thus faster and work without an internet connection.
}
