% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{ed_get}
\alias{ed_get}
\alias{get_ed_events}
\alias{get_ed_animals}
\alias{get_ed_specimens}
\alias{get_ed_tests}
\alias{get_ed_viruses}
\alias{get_ed_testspecimen}
\title{Functions to access main EIDITH tables}
\usage{
ed_get(endpoint, verbose = interactive(), postprocess = TRUE,
  header_only = FALSE, lmdate_from = "2000-01-01", lmdate_to = Sys.Date()
  + 1, auth = NULL, ...)

get_ed_events(verbose = interactive(), postprocess = TRUE,
  header_only = FALSE, lmdate_from = "2000-01-01", lmdate_to = Sys.Date()
  + 1, ...)

get_ed_animals(verbose = interactive(), postprocess = TRUE,
  header_only = FALSE, lmdate_from = "2000-01-01", lmdate_to = Sys.Date()
  + 1, ...)

get_ed_specimens(verbose = interactive(), postprocess = TRUE,
  header_only = FALSE, lmdate_from = "2000-01-01", lmdate_to = Sys.Date()
  + 1, ...)

get_ed_tests(verbose = interactive(), postprocess = TRUE,
  header_only = FALSE, lmdate_from = "2000-01-01", lmdate_to = Sys.Date()
  + 1, ...)

get_ed_viruses(verbose = interactive(), postprocess = TRUE,
  header_only = FALSE, lmdate_from = "2000-01-01", lmdate_to = Sys.Date()
  + 1, ...)

get_ed_testspecimen(verbose = interactive(), postprocess = TRUE,
  header_only = FALSE, lmdate_from = "2000-01-01", lmdate_to = Sys.Date()
  + 1, ...)
}
\arguments{
\item{endpoint}{Which API endpoint to access, one of "Event", "Animal",
"Specimen", "Test", or "Virus". Each endpoint delivers one of these tables.
Convenience functions are provided for each.}

\item{verbose}{Show a progress bar and other messages?}
}
\value{
a \link[tibble]{tibble}-style data frame
}
\description{
Functions to access main EIDITH tables
}

