% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{ed_get}
\alias{ed_animals}
\alias{ed_events}
\alias{ed_get}
\alias{ed_specimens}
\alias{ed_tests}
\alias{ed_viruses}
\title{Functions to access main EIDITH tables}
\usage{
ed_get(endpoint, username = NULL, password = NULL,
  verbose = interactive(), postprocess = TRUE, ...)

ed_events(username = NULL, password = NULL, verbose = interactive(),
  post_process = TRUE)

ed_animals(username = NULL, password = NULL, verbose = interactive(),
  post_process = TRUE)

ed_specimens(username = NULL, password = NULL, verbose = interactive(),
  post_process = TRUE)

ed_tests(username = NULL, password = NULL, verbose = interactive(),
  post_process = TRUE)

ed_viruses(username = NULL, password = NULL, verbose = interactive(),
  post_process = TRUE)
}
\arguments{
\item{endpoint}{Which API endpoint to access, one of "Event", "Animal",
"Specimen", "Test", or "Virus". Each endpoint delivers one of these tables.
Convenience functions are provided for each.}

\item{username}{Your EIDITH username. Store it in your \code{.Renviron} file
as \code{EIDITH_USERNAME=XXXXX} for automatic access.}

\item{username}{Your EIDITH password. Store it in your \code{.Renviron} file
as \code{EIDITH_PASSWORD=XXXXX} for automatic access.}

\item{progress}{Show a progress bar?}
}
\value{
a \link[tibble]{tibble}-style data frame
}
\description{
Functions to access main EIDITH tables
}

