% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{ed_metadata}
\alias{ed_metadata}
\title{EIDITH table metadata}
\usage{
ed_metadata()
}
\description{
Descriptions of fields in the EIDITH database.  Running \code{ed_metadata()} will
return a data frame with this information. \if{html}{This table is also shown below in a searchable and exportable format.}
}
\details{
More information can be found at the \href{https://eidith.org/Resources.aspx}{EIDITH resources page}.

\if{html}{

Here "Original Name" refers to the name in the EIDITH database, "Processed Name"
is what is found after running \code{\link{ed_process}} on the data tables.

\Sexpr[echo=FALSE, results=rd, stage=render]{tmp <- tempfile(fileext=".html");htmlwidgets::saveWidget(DT::datatable(eidith::ed_metadata()), tmp);return(paste(c('\\\out{<div>', readLines(tmp), '</div>}'), collapse = "\n"))}}
}

