% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{ed_process}
\alias{ed_process}
\title{Standard post-processing of EIDITH exports}
\usage{
ed_process(dat, endpoint)
}
\arguments{
\item{dat}{The data as exported from EIDITH and imported via the \code{\link[=ed_get]{ed_get()}} functions (without preprocessing).}

\item{endpoint}{The name of the API endpoint uaws: one of "Event",
"Animal",  "Specimen", "Test", "Virus", or "TestIDSpecimenID" (for test-specimen cross referencing).  Note these are different
than the names of the tables stored locally (which are lowercase and plural).}
}
\description{
This function takes raw data downloaded from EIDITH and puts it through
various preprocessing and cleaning steps.  In general there is no need to
call this function directly - it is called by both \code{\link[=ed_db_download]{ed_db_download()}} and the \link[=ed_get]{direct download functions}.
}
\details{
Steps taken to clean the data include:
\itemize{
\item Converting variable names from \code{camelCase} to \code{snake_case} to make it easy
to distinguish between raw and cleaned data.
\item Converting some variable names to clearer ones:  all \code{_id} variables are
numeric primary keys, other identifiers now go by \code{_id_name}.
\item Cleaning up whitespace and capitalization variability
\item Re-arranging table order to put the most pertinent information first.
\item Normalizing all animal taxonomic information to match the \href{https://www.itis.gov/}{ITIS} database.
\item Coercing some free-form entries (e.g. \code{specimen_type}) to a standard set of categories
\item Converting yes/no fields to TRUE/FALSE
\item Fixing spelling errors
\item Extracting common TRUE/FALSE variables from free-form text of viral interpretation (Genbank numbers, whether virus is known, whether virus is known to affect humans).
}
}
