% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{eidith_user}
\alias{eidith_pwd}
\alias{eidith_user}
\title{EIDITH logon credentials}
\usage{
eidith_user(verbose = interactive())

eidith_pwd(verbose = interactive())
}
\arguments{
\item{verbose}{Show messages?}
}
\description{
These are convience functions for using your default EIDITH logon credentials
for any calls to EIDITH.  Place your EIDITH username and password in your
home \code{.Renviron} file as \code{EIDITH_USERNAME} and
\code{EIDITH_PASSWORD} and these values will be used as defaults for all
calls.
}

