% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{p2_api_endpoints}
\alias{p2_api_endpoints}
\alias{p1_api_endpoints}
\title{This function returns the names of the tables in the PREDICT-2 EIDITH database.}
\usage{
p2_api_endpoints()
}
\description{
This function returns the names of the tables in the PREDICT-2 EIDITH database.
}
