% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{p1_api_endpoints}
\alias{p1_api_endpoints}
\alias{p2_api_endpoints}
\alias{p2_api_endpoints}
\alias{p1_api_endpoints}
\title{Return he names of the tables in the EIDITH database.}
\usage{
p1_api_endpoints()

p2_api_endpoints()
}
\description{
Return he names of the tables in the EIDITH database.

This function returns the names of the tables in the PREDICT-2 EIDITH database.
}
