% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_local_database.R
\name{import_local_db}
\alias{import_local_db}
\title{Downloads EIDITH database from google drive}
\usage{
import_local_db(database = c("eha", "eha_with_malaysia", "global"))
}
\arguments{
\item{database}{whether downloading "eha", "eha_and_malaysia",or "glabal" database}
}
\description{
Note that post-September 2020, EIDITH remote database access is disabled.
}
\details{
This function downloads a static version of the database. Access is enforced via Google Drive.
You will be prompted in your R console to allow the \code{googledrive} package to access to your Google account.
See \code{?googledrive::drive_auth()}.

Contact Emma (mendelsohn@ecohealthalliance.org) for access or with any questions!
}
