% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_local_database.R
\name{import_local_db}
\alias{import_local_db}
\title{Downloads EIDITH from google drive}
\usage{
import_local_db(database = c("global", "eha", "eha_and_malaysia"))
}
\arguments{
\item{database}{whether downloading "global", "eha", or "eha_and_malaysia" database (all password protected)}
}
\description{
Downloads EIDITH from google drive
}
